Implementation of simple boolean operators (bug!)
This completes the basic implementation of boolean operators
(the simple ones that eagerly evaluate - and and or will need
to wait for control flow!)

That said, I have a bug in my parser; the example program
`!(5 - 4 > 3 * 2 == !nil)` from the end of the chapter is not
compiling; some poking around suggests that when I have binary
operators inside groupings something goes wrong and I never find
the right parenthesis.

I'll need to debug this before going on to Chapter 19.