Fix bug in Pratt parsing
I think the book may actually have an error here, I don't recall
the unary operator actually having a dynamic precedence anywhere
in the text.

But to behave correctly, we can't use PREC_UNARY - the NOT operator
is supposed to have loose binding (or at least I thought so? It
certainly does in most languages).

To make this easy to find, I added a PRINT_DEBUG macro and quite
a bit of debugging so that we can trace the core pratt parser loop.

Once I found the bug, I also extracted the bump-precedence-and-
call-parse-precedence hook into its own function named for the
actual action we're taking in context, which is parsing the RHS
of a (unary *or* binary) operator.

This fixes the bug.