Add string equality, fix another parser bug
I manually copied the Pratt table, and just realized that I
forgot the TOKEN_ERROR and TOKEN_EOF rules; as a result I was
segfaulting sometimes after looking up TOKEN_EOF in the table.

My guess is that I hadn't noticed yet because most of the time
my out-of-bounds reads happened to return values that would break
me out of the parsing loop, but this was nondeterministic and now
that I have strings it's easier to allocate more memory and hit
the undefined behavior.