Finish globals: setting global variables
This makes the compiler code noticeably messier.

I think it's a good example of where a single-pass compiler can
start to hit its limits and get ugly; the jlox approach of building
an expression and then verifiying that it was a legal target was
easier to understand than setting state on a recursive loop.

Finished with Chapter 21, ready for Chapter 22 on local vars.