Add local variables
This was fun.

I decided to diverge from clox in the semantics of code like this:
```
var x = 1;
{
  var x = x + 1;
}
```

In the book's clox this is illegal. I decided to make it legal;
inside the initializer to the inner `x` the name `x` still resolves
to the outer `x`.

This feels a little more natural to me, probably because of ocaml
where redefines are scoped this way.