Start working on functions
Currently ready to start 24.4 on defining
functions.

The interpreter is now entirely frame based:
- the vm stack is still global
- but the chunk is now function-local:
  - the vm holds onto an "active function" frame
  - the ip and a frame pointer into the global stack live there
  - the active function also lives there
    - the chunk is accessed via this

We also push the active function (at this point just the top-level
since we cannot define functions yet) onto the stack before we start
(this is why there's now a placeholder local, to ensure that local
stack offsets account for this extra entry on the stack).

As far as I can tell the interpreter works correctly; we don't have
functions yet but the outer compile and vm loops have been
"functionalized" so to speak.