Wire up closure code.
The vm frame now tracks a closure rather than a function
(at this point the closure is just a pointer, so all previous
code works as before with an extra level of indirect).

Instead of pushing a function directly to the stack in function
definitions, we now push a closure using a dedicated OP_CLOSURE.

We now have a layer where we can add the variable capture logic.

Ready to start Section 25.2 on Upvalues
https://craftinginterpreters.com/closures.html#upvalues