Closures now work!
I have a nice demo program. It's funny, debugging the demo took
me almost as long as debugging the vm itself haha.

My version of `captureUpvalue` is a little cleaner structurally,
at the cost of more complex pointer arithmetic around insertion.
I'm kind of happy with it as an example of double-pointer fiddling.