Fix last known GC bug
The issue here was that I was allocating the closure before
allocating the upvalues array - so allocating the upvalues would
garbage collect my closure in the middle of initialization, before
I even had a final value that I could put in the stack!

The fix is to allocate upvalues (which is just an array of null
pointers, the actual upvalues get created later so there's no GC safety
concern) *before* I allocate the closure.

Once again, if I'd copy-pasted from the book this would not
have come up haha!

---

It's probably generally a good idea to allocate objects as late
in construction as possible. That said, it's worth noting that there
are a number of alternative tricks we could use:

- in this particular case, it would be easy enough to just call
  malloc directly. Although to some extent that would defeat the
  purprose of GC, which is to clean up if things are getting big!
  As a rule we only want to use raw malloc / realloc in the GC
  itself, and have all other allocations pass through our
  potentially-GC-triggering code
- instead of pushing to the vm stack, we could track a separate
  stack of just-allocated objects inside the object code and use
  that as an extra gc root for "temp" values.

----

Note that I still haven't implemented the tuned GC logic. I'm not
really that interested at the moment so I'll leave that as an
exercise for later :)